[VTA] Fix vta rpc server, refactor launch cond to not depend on sys.argv (#8671)
